+++ test/transform/to_camel.js
@@ -11,14 +11,14 @@ const MODULE_REQUIRE = 1
     , leca = noda.inRequire('index')
     ;
 
-describe.only('Used as transformer', () => {
+describe('Used as transformer', () => {
     it('toCamel', () => {
         let ci = new leca.Case({
 			splitter: /[\s_\-]/,
 			jointer: '-',
             wordFormatter: word => word.toLowerCase(),
         }); 
-    
+	
         assert.equal('foo-bar', ci.reformat('foo BAR'));
 	});
 });