+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning 2.0.0](http://semver.org/).
 
+##	[0.0.8] - Feb 15th, 2018
+
+*	Predefined case `leca.kebab` added.
+*	Predefined case `leca.pascal` added.
+*	Predefined case creator `leca.kebab.terms()` added.
+*	Predefined case creator `leca.pascal.terms()` added.
+*	Restrict `leca.camel.parse()`  to return false when the first character is upper case.
+
 ##	[0.0.7] - Feb 14th, 2018
 
 *	Make `<case>.format()` to accept seperated string arguments beyond a string array.